refactor: 说话者轴收成具名类型 ArchiveRole，四套数字词表从此同形
Why: 宿主每轮都要声明「谁说的」，而 role 是四套数字词表里唯一没有名字的一条——门面直接发布裸
uint8，宿主把读回的行渲染进 prompt 时得自己写 0/1/2/3 那张表，指南也得专门记一句例外。

- core 新增 ArchiveRole（名字表 + String() + Valid()），ArchiveSlot / SceneMessage /
  api.ArchiveInput / api.ArchiveSlot 四处字段同改，常量原样赋值不经换算
- content.ValidateAppend 把两问分开答：数字在不在词表里判 Valid()，「这个词轮不轮到调用方写」
  仍点名 user/agent/system，于是凭空发明的数字与想冒充巩固产物的 RoleDream 各有各的拒绝理由
- 负例把自己两条门禁推翻并各自处置：反射检查「形状字段类型等于常量类型」在真实回归上永不为红
  （两侧同为 uint8 时自动成立，退回裸 uint8 是编译失败），删；整篇文档搜取值词的指南检查假绿
  （dream 还活在 memory_dream 与 internal/dream 里），改为按 enum-wire-table 标记逐行核，
  三个负例各自红在被改的那一行，两套字符串词表一并盖住
- internal/agent.md 那句「公开常量里没有 RoleDream」按实测改口（门面一直导出它）

+++ api/surface_closed_test.go
@@ -37,7 +37,7 @@ func TestEveryCallAnswersErrClosedAfterClose(t *testing.T) {
 	}
 	sceneID, topicID := searched.Scene.SceneID, searched.NewTopicID
 	if _, err := sess.AppendArchive(ArchiveInput{
-		Kind: KindUtterance, Role: uint8(RoleUser), ContentType: ContentText,
+		Kind: KindUtterance, Role: RoleUser, ContentType: ContentText,
 		Content: "hello", CreatedAt: 1_770_000_000_000,
 	}); err != nil {
 		t.Fatalf("AppendArchive: %v", err)
